Move current location to button, use google/apple maps uri
